Move stress calculations to ADV directory
