+++ changelog.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unversioned
+    - Bugfixes
+	    - "calc_tke" function now errors if given multi-beam ADCP velocity vector
+	    - Clean thresh function failed to remove spikes
+		- ADV despiking max gap input
+		- Fix input error to calc_L_int
+
+	- API/Useability
+		- Moved "calc_stress" and "calc_csd" functions to ADVBinner class
+		- Removed "do_tke" function
+
 ## Version 1.0.2
     - Bugfixes
 	    - Detect missing timestamps in ADV binary files
