+++ dolfyn/adv/turbulence.py
@@ -87,6 +87,119 @@ class ADVBinner(VelBinner):
 
         return out
 
+    def calc_stress(self, veldat, detrend=True):
+        """Calculate the stresses (cross-covariances of u,v,w)
+
+        Parameters
+        ----------
+        veldat : xr.DataArray
+            Velocity data array from ADV data. The last dimension is 
+            assumed to be time.
+        detrend : bool (default: True)
+            detrend the velocity data (True), or simply de-mean it
+            (False), prior to computing stress. Note: the psd routines
+            use detrend, so if you want to have the same amount of
+            variance here as there use ``detrend=True``.
+
+        Returns
+        -------
+        ds : xarray.DataArray
+
+        """
+        time = self.mean(veldat.time.values)
+        vel = veldat.values
+
+        # Will error for ADCP 4-beam
+        out = np.empty(self._outshape(vel.shape)[:-1],
+                       dtype=np.float32)
+
+        if detrend:
+            vel = self.detrend(vel)
+        else:
+            vel = self.demean(vel)
+
+        for idx, p in enumerate(self._cross_pairs):
+            out[idx] = np.nanmean(vel[p[0]] * vel[p[1]],
+                                  -1, dtype=np.float64
+                                  ).astype(np.float32)
+
+        da = xr.DataArray(out, name='stress',
+                          dims=veldat.dims,
+                          attrs={'units': 'm^2/^2'})
+        da = da.rename({'dir': 'tau'})
+        da = da.assign_coords({'tau': ["upvp_", "upwp_", "vpwp_"],
+                               'time': time})
+        return da
+
+    def calc_csd(self, veldat,
+                 freq_units='Hz',
+                 fs=None,
+                 window='hann',
+                 n_bin=None,
+                 n_fft_coh=None):
+        """Calculate the cross-spectral density of velocity components.
+
+        Parameters
+        ----------
+        veldat   : xarray.DataArray
+          The raw 3D velocity data.
+        freq_units : string
+          Frequency units of the returned spectra in either Hz or rad/s 
+          (`f` or :math:`\\omega`)
+        fs : float (optional)
+          The sample rate (default: from the binner).
+        window : string or array
+          Specify the window function.
+        n_bin : int (optional)
+          The bin-size (default: from the binner).
+        n_fft_coh : int (optional)
+          The fft size (default: n_fft_coh from the binner).
+
+        Returns
+        -------
+        csd : xarray.DataArray (3, M, N_FFT)
+          The first-dimension of the cross-spectrum is the three
+          different cross-spectra: 'uv', 'uw', 'vw'.
+
+        """
+        fs = self._parse_fs(fs)
+        n_fft = self._parse_nfft_coh(n_fft_coh)
+        time = self.mean(veldat.time.values)
+        veldat = veldat.values
+
+        out = np.empty(self._outshape_fft(veldat[:3].shape, n_fft=n_fft),
+                       dtype='complex')
+
+        # Create frequency vector, also checks whether using f or omega
+        coh_freq = self.calc_freq(units=freq_units, coh=True)
+        if 'rad' in freq_units:
+            fs = 2*np.pi*fs
+            freq_units = 'rad/s'
+            units = 'm^2/s/rad'
+            f_key = 'omega'
+        else:
+            freq_units = 'Hz'
+            units = 'm^2/s^2/Hz'
+            f_key = 'f'
+
+        for ip, ipair in enumerate(self._cross_pairs):
+            out[ip] = self._cpsd(veldat[ipair[0]],
+                                 veldat[ipair[1]],
+                                 n_bin=n_bin,
+                                 n_fft=n_fft,
+                                 window=window)
+
+        da = xr.DataArray(out,
+                          name='csd',
+                          coords={'C': ['Cxy', 'Cxz', 'Cyz'],
+                                  'time': time,
+                                  f_key: coh_freq},
+                          dims=['C', 'time', f_key],
+                          attrs={'units': units, 'n_fft_coh': n_fft})
+        da[f_key].attrs['units'] = freq_units
+
+        return da
+
     def calc_epsilon_LT83(self, psd, U_mag, omega_range=[6.28, 12.57]):
         """
         Calculate the dissipation rate from the PSD
@@ -145,7 +258,7 @@ class ADVBinner(VelBinner):
         Parameters
         ----------
         vel_raw : xarray.DataArray
-          The raw velocity data (with dimension time) upon 
+          The raw velocity data (1D dimension time) upon 
           which to perform the SF technique. 
         U_mag : xarray.DataArray
           The bin-averaged horizontal velocity (from dataset shortcut)
@@ -163,6 +276,8 @@ class ADVBinner(VelBinner):
 
         """
         veldat = vel_raw.values
+        if len(veldat.shape) > 1:
+            raise Exception("Function input should be a 1D velocity vector")
 
         fs = self._parse_fs(fs)
         if freq_rng[1] > fs:
@@ -319,8 +434,8 @@ class ADVBinner(VelBinner):
         scale = np.argmin((acov/acov[..., :1]) > (1/np.e), axis=-1)
         L_int = U_mag.values / fs * scale
 
-        return xr.DataArray(L_int, name='L_int', 
-                            coords={'dir':a_cov.dir, 'time':a_cov.time}, 
+        return xr.DataArray(L_int, name='L_int',
+                            coords={'dir': a_cov.dir, 'time': a_cov.time},
                             attrs={'units': 'm'})
