+++ dolfyn/tests/data/BenchFile01_func.nc
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:5118120805f236ac918a37916f72edda0895fbc34cc92c56adc03c6af7611706
-size 16419
+oid sha256:76663743e9a71772d3bd92b831048555d98aefe4a9ff2873fe3f90413758e21c
+size 10245
