+++ dolfyn/tests/data/vector_data01_avg.nc
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:3594722ae19da96fffdbc9a1d86af7ae8a3f8754db8ce54e182d06cc4248459a
-size 37441
+oid sha256:a483de8b13985873a8d30f93093c7a3045b77104ef7e93466b35332068e67f24
+size 35281
