+++ dolfyn/tests/data/vector_data01_func.nc
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b45c861f31749731cd0bedd5e15ea2f3010096130826014128653404c22cc168
-size 22311
+oid sha256:38063e27c304a98a484ea154920889926dde3dd31f1ee47e877ea108d4564c9a
+size 16153
